Allow overriding the system clock fixes #118
Adds an implicit clock to the time based methods in JwtTime, and adds an
implicit val clock to JwtCore. Give each Jwt object a companion class
that can be constructed with a different Clock, for users that wish to
use a different clock than the system one. Remove MockIt from tests and
just use overriden clocks instead.

Since JwtSession calls out to JwtJson, added a helper method that will
look at the Clock being used, and if it is referentially equal to the
Clock being used by the JwtJson object, it will use the object instead
of constructing a new JwtJson.