Use EdDSA instead of Ed25519
Fixes #464.

`Ed25519` is not a valid
[Json Web Algorithm](https://www.rfc-editor.org/rfc/rfc7518#section-6.2) name.
It is the name of a curve used when the `EdDSA` algorithm is in use, and in
that instance, `EdDSA` is the correct algorithm name to use. `Ed25519` is just
a property of the private key, alternate curves can also be used, such as
`Ed448`, the use of these curves don't impact the algorithm being used.